Add files via upload
This is sample data to create SQL tables.The tables can be created by running SQL scripts . Also there is demo data for populating the tables. These are also SQL scripts. Be sure when creating schema to name schema as drone_inventory. Username must be "root" and password is 1234. If you change the Username,password, and or schema....the changes must be hard coded in the servlets.